requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
ultralytics
pillow
fpdf2
